services/client-api/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Validation (use shared schemas)
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Redis (for caching/sessions)
redis>=5.0.0

# HTTP client (for ElevenLabs API)
httpx>=0.26.0

# Utilities
python-dotenv>=1.0.0
